refactor: remove RequestBulkDeleteEntity and related tests, introduce BulkCleanupEntity
- Deleted RequestBulkDeleteEntity.spec.ts and associated use case.
- Introduced BulkCleanupEntityUseCase and its factory.
- Updated EntityPermissionChecker to return templateId along with sharedId.
- Modified MongoEntityPermissionChecker to handle new output structure.
- Created BulkCleanupEntityJob to handle bulk cleanup operations.
- Updated BulkDeleteEntityController to use new BulkCleanupEntityUseCase.
- Implemented deleteReferencesToSharedIds method in MultiLanguageEntityDataSource.
- Adjusted relationships data source to support bulk deletion by sharedId.
- Enhanced search functionality to refresh after bulk operations.
- Updated queue registry to register new BulkCleanupEntityJob.
