+++ app/api/config.ts
@@ -89,7 +89,7 @@ export const config = {
       fileCacheHeaders: FEATURE_FLAG_FILE_CACHE_HEADERS === 'true' || false,
       v2UploadFile: false,
       v2CreateEntity: false,
-      v2BulkDeleteEntity: false,
+      v2BulkDeleteEntity: true,
       v2CSVImport: false,
     },
   },
