+++ app/api/entities.v2/contracts/MultiLanguageEntitiesDataSource.ts
@@ -9,6 +9,9 @@ export interface MultiLanguageEntityDataSource {
 
   deleteMetadataProperties(propertyNames: string[], sharedIds: string[]): Promise<void>;
   bulkDelete(sharedIds: string[]): Promise<void>;
+  deleteReferencesToSharedIds(
+    entities: Array<{ sharedId: string; templateId: string }>
+  ): Promise<void>;
 
   renameMetadataProperties(
     propertyNames: { [oldName: string]: string },
