+++ app/api/search/search.js
@@ -895,6 +895,7 @@ const search = {
       },
       conflicts: 'proceed',
       wait_for_completion: true,
+      refresh: true,
     });
   },
